refactored with package worker
